docs: Clarify use of KERNELHEADERS
In the Makefile we specify additional levels of indirection beyond that
of the normal kernel module Makefile.  In this case we have added both
KERNELVER and KERNELHEADERS variables so that they can be supplied at
runtime as command line arguments.  The process for doing this is now in
README.md.